+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## master (unreleased)
 
+- Fix background migrations `#progress` possibility to fail with zero division error
 - Add `ensure_background_data_migration_succeeded` and `ensure_background_schema_migration_succeeded` migration helpers
 - Raise in development when background index creation/removal was not enqueued
 - Suggest two migrations for adding foreign keys